ageisai-sdk/requirements.txt
requests
google-generativeai
scikit-learn
